Deleted nested repository
