+++ DualInputs.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+
+def dual_dictionary():
+    print('interleaved')
