+++ main.py
@@ -1,0 +1,5 @@
+from shapes import Circle
+
+circle_instance = Circle(5)
+
+print(circle_instance.area())
